8260019: Move some Thread subtypes out of thread.hpp
