+++ src/hotspot/share/ci/ciEnv.hpp
@@ -31,6 +31,7 @@
 #include "code/debugInfoRec.hpp"
 #include "code/dependencies.hpp"
 #include "code/exceptionHandlerTable.hpp"
+#include "compiler/compilerThread.hpp"
 #include "compiler/oopMap.hpp"
 #include "oops/methodData.hpp"
 #include "runtime/thread.hpp"
